modem: reset cancellable after cancelling dbus methods
I first wanted to use "reset" method, but according to [1]:
"The recommended practice is to drop the reference to a
cancellable after cancelling it, and let it die with the outstanding
async operations. You should create a fresh cancellable for further
async operations."

Pretty weird that in other parts of documentation it still says that it
can be used for concurrent events.

[1] https://developer.gnome.org/gio/2.54/GCancellable.html#g-cancellable-reset

Change-Id: I7392275620576fd0037be5b9d9f360a46a4f8312