OsmoGSMTester: Installation: change some confusing terminology
It's difficult to understand when it referrs to jenkins job in general
and when it talks about specific jenkins jobs that builds binaries required for
osmo-gsm-tester and the ones which run osmo-gsm-tester.

As a result, it's difficult to understand which sections apply to each
host in case more than one host is used.

Change-Id: I5c882e77b469629028b8d773053783e18f3d1737