example: Maintain different resources.conf for RnD and prod
We still want to maintain this file in the same osmo-gsm-tester repo
because we frequently neef to update the config when adding new
features.

Until now only 1 file was maintained (which was used for RnD setup), and
then when runnin in prod the jenkins script used sed to change the file
to accomodate slightly changes. This way is too hacky, so let's just
maintain too separate files, keeping the original resources.conf key
name used by osmo-gsm-tester free, so that jenkins job can symlink one
of the 2 files to it.

Take the chance to remove OctoBTS and Sysmocell5k from the RnD resources
file, as we don't have those them.

Change-Id: Ifec851c7ac6fca6b294e57dfe86b92f214ae8f42