jenkins-build-amarisoft: fix build script for ZMQ module
newer versions of srsLTE create symlinks with the rf library
including the version, see below. This caused the trx_zmq.so
lib to load .so.0 instead of .so only. this patch adds
another symlink so loading works fine again

libsrslte_rf.so -> libsrslte_rf.so.0
libsrslte_rf.so.0 -> libsrslte_rf.so.20.10.0
libsrslte_rf.so.20.10.0

Change-Id: Ib5302c9872f4c82a1ee5074f59de86709aa6c5a9