amarisoft_mme: use/expose count paramter from MS config for MME
the UEs count param, if specified, is mapped into the config
values of the EPC so it can be used for template rendering.

Since we only have one count parameter, its added to the first
subscriber. 1 is the default value and its used when no count
param is specified.

this allows to run the Amarisoft UE simulator with the Amarisoft EPC
without having to create subscriber entries for each simulated UE.

Change-Id: I29885791f716c204d0b6f18ba134885bae853b6f