jenkins-build-common: adds support to the new variables for patchelf.
2 variables are added in order to increase the flexibility of the add_rpath function.
patchelf_rpath_extra_args: adds more arguments to the rpath patch. For example "--force-rpath"
patchelf_rpath_dir: path to the libraries. This path is appended to the default '$ORIGIN/../lib/'

This commit add the jenkins-build-srslte-bbu.sh script, that builds the given repository filling the information for the 2 newly added variables.

Change-Id: I9f510112c63e7598add8c00f8573e34cac1b6064